requirements.txt
Django==3.2.*
django-ninja==0.13.0
django-extensions==3.1.3
python-decouple==3.4
django-localflavor==3.0.1
pytest-django==4.2.0
djoser==2.1.0
django-cors-headers==3.7.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
dj-static==0.0.6
